+++ include/crsum.hpp
@@ -4,8 +4,8 @@
 
 class CRsum final: public CRobj {
     public: 
-        CRsum(int i, size_t l); 
-        CRsum(int i, double x, double h);
+        CRsum(long long i, size_t l); 
+        CRsum(long long i, double x, double h);
         
         std::unique_ptr<CRobj> add(const CRobj& target) const override;
         std::unique_ptr<CRobj> mul(const CRobj& target) const override;
@@ -22,11 +22,11 @@ class CRsum final: public CRobj {
         void print_tree() const override;
         void simplify() override;
         
-        std::string genCode(size_t parent, size_t index, int place,std::string indent) const override;
+        std::string genCode(size_t parent, long long index, long long place,std::string indent) const override;
         
         std::unique_ptr<CRobj> copy() const override;
 
-        inline void shift(size_t i ) noexcept override final {
+        inline void shift(long long i ) noexcept override final {
             if (index > i){
                 for (size_t j = 0; j < isanumber.size(); j++){ 
                     operands[isanumber[j]]->shift(i);
